back-end/requirements.txt
fastapi[standard]
uvicorn[standard]
pyswisseph
python-dateutil
kerykeion
pytz
timezonefinder
